Initial commit
Added files, as far as it goes. For now, some crude windowing and rendering
capabilities have been implemented, as well as the basic structure required to
add more sophisticated mechanisms such as buttons, labels, etc.

I believe that in the future, I am going to have to make assumptions about
things, e.g. names and/or locations of font files. I might, for instance, add a
static method to GraphicsHelper which loads the first font in a given list, and
let the SDLBox window handle loading one of these fonts as the "default font"
upon init. While I'm at it, maybe I should rename GraphicsHelper to something
more generic, such as ResourceHelper (or ResourceLoader) or something. We'll
see.