Implement panels
Panels are sdlbox Components which can contain other sdlbox Components. They
have an orientation (vertical or horizontal), dimensions corresponding to the
total width and height of the contents, and, uh, stuff. Drawing them draws all
other components to the screen.

Note that panels still need to be created with the new keyword before added to
the window - this is to avoid a segfault due to a double delete (manual delete
in window in addition to automatic delete as the stack unloads).

Furthermore, the "set" methods of Component have been altered, and are now
"with" methods which return a pointer to the object in question (this).