Major update
Basically introduced:
- "Textual" (abstract) components (for instance labels)
- "TextureComponent" which have Texture pointers
- "VolatileLabel" which are expected to change its content frequently and thus
  [the Texture] is recreated each draw (normal labels are recreated every update)
- Custom user events
- Lots of primitive shape drawing
- More test programs
- Static library libsdlbox.a
- Bunch of other shit, probably

TODO: I'm planning to give SDLBox built-in support for "rooms". It will be able
to "go to" a room, the add method will add stuff to the current room, and the
destructor will destroy the contents of each room before exiting.

Planned classes:
- Rooms (keeps track of legal rooms)
  - void addRoom(string roomName)
  - Room getRoom(string roomName)
- Room (basically a glorified ID)
  - Room(int id) (used only by Rooms - friend constructor?)
  - Room(string roomName) (get existing room with name roomName, or exception)
  - operator ==(string roomName) (compare to existing room with same name)
  - operator ==(Room other) (compare room id)
  - ... + maybe some more

Planned SDLBox methods:
- goToRoom(Room room) (change currently active room)
- wipe() (destroy and remove contents of currently active room)
- wipe(Room room) (destroy and remove contents of specified room)
- ... + probably something else I think of

Planned SDLBox modified methods:
- add: adds to current room only
- dtor: destroy contents of all rooms
- ... + probably some more

NOTE: Rooms will likely not be supported unless a fixed size is given upon
window construction - this is in order to simplify a lot of things, so I don't
get in trouble.