Add Room and Rooms to makefile, add activeRoom
Created an activeRoom private variable in SDLBox. To that end, created an empty
constructor in Room, which sets id to -1 to avoid confusing it with a legitimate
room. If I add enough rooms for overflow to become a concern in this regard,
I'll have bigger problems anyways.

Also added Room and Rooms to makefile objects to avoid "undefined reference"
linker error upon build.