Add score label, fix jumping texture on setText
I added a score label. Upon doing this, I found that setText made the label
suddenly jump to 0, 0. It turns out this happened because using setText creates
a new texture overriding the old one, but this is never repositioned (unless, of
course, it is affected by repositionChildren, which it's not when
receivePosition is false). So I added a call to repositionTexture in the
setTexture method in TextureComponent, and it worked like a charm.