Remove dependencies on deprecated methods
I removed the calls to "with<attribute>" methods scattered throughout the code
base and replaced them with calls to the ComponentFactory API, which can also be
used for purposes like this even though it's not quite a factory anymore at this
rate. This is horrible practice, never do this.

At any rate, a gazillion lines of debug print statements added and removed
later, here we are, and the dependencies on the deprecated methods should be
gone. I introduced a "commit" method instead (already present in Label, but now
available in all Components) which propagates changes affecting Component down
to its subclasses. By default, this method does nothing. This means I can
override the commit method instead of the withPosition methods in order to
reposition the Texture internally present in Label and whatnot - the information
necessary is already there, I'm just propagating it downwards in the inheritance
hierarchy.

On a side note, my git log looks pretty fucked up now, because I thought I'd be
clever and instead managed to screw myself over with git stash and some detached
head or whatever. After failing to properly rebase several times, I ended up
accepting my defeat and move on. So if anyone wonders why I have duplicate
commits, that's why. I'm too scared to try and fix it now.