Add dispatchEvent function
Sometimes we just care that "an event happened", rather than the specific
details associated with it. This function allows us to quickly and painlessly
dispatching an event of a certain type, ignoring all details. It does not help
in dispatching more elaborate events, but in the case that this is needed,
doing it in a one-liner might not be of paramount concern.