Set position handling data type to double
This is in order to have more fine-grained control over changing positions in
any Component - position updates don't necessarily need to be performed entire
pixels at a time, although the external representation remains integer-based.

TODO: Update ComponentFactory to take doubles in order to reflect this change.

NOTE: I might also want to add a changePosition protected method to Component
(or a changeXPos and a changeYPos method) which adds the specified amount to the
position, evading the need to go through ComponentFactory at all. Or I could
make this a method available in ComponentFactory, although I'd be forever
relinquishing my interpretation of it as a "factory" at that point.