Add preliminary collision event capabilities
I'm letting SDLBox detect collisions in the step method. Collision detection
works as follows:
- For each component, if component is collidable, add to "collidables" vector.
- For each component c in collidables, loop through each subsequent component.
- For each subsequent component d, if c collides with d, push a collision event.
  - Collision events have data1 set to c and data2 set to d.

Checking whether an object is collidable is left up to the Component's
"collidable" method. By default, a Component is not collidable, but this method
can be overridden in subclasses which should interact with each other through
collisions. The actions performed upon a collision should be left up to the
event listener.

NOTE: Even though it's listed as a subfeature in the Github issue for collision,
collision classes is not a priority until I try to develop something which
requires it (or something for which it would be handy). For now, I'll push the
responsibility of determining whether a collision is applicable down to the event
listeners.