Attempt #1 at fixing #5 - perpetual reposition
Issue #5: bug causes segfaults in earlier test programs (programs with
automatically readjusting positions). Turns out this is because a line in
Textual which calls SDLBox::repositionChildren whenever it commits - this is a
problem, because this calls the reposition method of the component itself, which
calls the component's commit, which calls SDLBox::repositionChildren....

The fix: Don't allow calling repositionChildren directly from a commit. I made a
new method, scheduleReposition, which sets a flag in SDLBox. If that flag is set
when the step method rolls around, SDLBox ITSELF calls the repositionChildren
method (which probably results in the flag being set multiple more times, but
that does not matter because it's unset after repositionChildren is done
executing anyways). This will probably make Textual components way more
efficient, too, since they don't reposition everything, all the time.

Also, repositionChildren was made private. Finally. This was long overdue, I
can't believe I didn't do this before ._.