Fix #5 - perpetual reposition
All old example programs have been tested with the updated code, are compiling
and are working. In addition, the panelTest program has been modified because
for some reason, the way it was made all the components more or less take up no
space at all...? I'm not sure why or what's going on. This marks the closeure of
the perpetual reposition bug.

TODO: Open a new issue, about the Panel thing.